+++ peakrack_kyc/lib/Providers/AlipayFaceProvider.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace PeakRack\Kyc\Providers;
+
+class AlipayFaceProvider extends ReservedProvider
+{
+    public function getName(): string
+    {
+        return 'alipay_face';
+    }
+}