+++ peakrack_kyc/lib/Providers/CompanyVerificationProvider.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace PeakRack\Kyc\Providers;
+
+class CompanyVerificationProvider extends ReservedProvider
+{
+    public function getName(): string
+    {
+        return 'company_verification';
+    }
+}