+++ peakrack_kyc/lib/Providers/OverseasKycProvider.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace PeakRack\Kyc\Providers;
+
+class OverseasKycProvider extends ReservedProvider
+{
+    public function getName(): string
+    {
+        return 'overseas_kyc';
+    }
+}